fixed the XID problem.
git-svn-id: https://r-osgi.svn.sourceforge.net/svnroot/r-osgi/trunk@1178 c960cec6-6b13-0410-98af-d8f80e9b129c

+++ transport/mina/src/main/java/ch/ethz/iks/r_osgi/transport/mina/codec/RemoteOSGiMessageCodec.java
@@ -60,7 +60,7 @@ public abstract class RemoteOSGiMessageCodec implements MessageEncoder,
 		if (funcID != type) {
 			return MessageDecoderResult.NOT_OK;
 		}
-		final Short xid = new Short(in.getShort());
+		final Integer xid = new Integer(in.getInt());
 		session.setAttribute("xid", xid);
 
 		final int length = in.getInt();
@@ -103,7 +103,7 @@ public abstract class RemoteOSGiMessageCodec implements MessageEncoder,
 
 		buf.put((byte) 1); // version
 		buf.putShort(type); // funcID
-		buf.putShort(msg.getXID()); // xid
+		buf.putInt(msg.getXID()); // xid
 
 		buf.skip(4); // skip length
